fix uint error

+++ ConsoleNFT_Source_Data.sol
@@ -8,13 +8,13 @@ contract ConsoleNFT_Source_Data {
     uint16 epic = 25; // 5%
     uint16 rare = 50; // 10%
     uint16 uncommon = 150; // 30%
-	// rest = 54%
+    // rest = 54%
 	
-	mapping(uint16 => uint8) private _vaultRarities;
+    mapping(uint => uint8) private _vaultRarities;
 	
 	constructor() {
 		
-		uint16[150] memory list;
+	uint16[150] memory list;
 
         list = [210,196,66,163,222,255,54,241,426,218,380,156,470,385,2,340,266,366,478,262,30,164,472,120,394,35,294,68,422,90,180,170,0,458,404,315,270,497,11,334,29,309,162,329,6,275,58,459,286,57,32,232,18,384,152,160,114,451,496,282,310,209,108,320,302,36,408,159,174,141,240,288,226,72,234,360,438,462,21,289,450,219,244,50,158,25,228,28,454,498,20,337,77,299,212,390,418,12,76,110,99,59,24,147,354,215,144,285,52,343,83,447,146,89,138,60,414,149,80,81,130,300,101,444,92,100,38,325,4,73,410,8,17,15,348,400,132,387,230,263,95,161,358,126,338,305,186,313,128,123];
 		
@@ -42,7 +42,7 @@ contract ConsoleNFT_Source_Data {
 		
 	}
 	
-    function getVaultRarity(uint16 token_id) public view returns(uint8) {
+    function getVaultRarity(uint token_id) public view returns(uint8) {
 
         return _vaultRarities[token_id];
         
@@ -126,6 +126,5 @@ contract ConsoleNFT_Source_Data {
         return output;
 
     }
-	
-	
+    
 }
